list and sets
Write a Python program to read two strings of words separated by a comma. Convert each string
to a set of unique words. Then print out a set containing all the words that occur in the two sets together.
Print out a second set that contains the words that occur in both sets. Assume that the user may enter
spaces as well. Therefore, you will need to follow the steps given below:
• Read in the string
• Separate it into a list using the ‘,’ as a delimiter
• Use list comprehension to strip each word of any leading and trailing spaces
• Convert each list to a set
• Use appropriate set methods to print out the required output